Refactored base sql task class
